AUG/projekt/requirements.txt
requests
antlr4-python3-runtime